fix circle type dependency
