+++ ui/src/context/CustomComponentContext.tsx
@@ -1,14 +1,20 @@
 import React, { createContext, ReactNode, useMemo } from 'react';
-import { CustomElementsMap } from '../types/CustomTypes';
+import { CustomControlConstructor } from '../components/CustomControl/CustomControlBase';
+import { CustomTabConstructor } from '../components/CustomTab/CustomTabBase';
+import { CustomHookConstructor } from '../types/components/CustomHookClass';
 
-const CustomComponentContext = createContext<CustomElementsMap | undefined>(undefined);
+export type CustomComponentContextType =
+    | Record<string, CustomHookConstructor | CustomControlConstructor | CustomTabConstructor>
+    | undefined;
+
+const CustomComponentContext = createContext<CustomComponentContextType>(undefined);
 
 export function CustomComponentContextProvider({
     children,
     customComponents,
 }: {
     children: ReactNode;
-    customComponents?: CustomElementsMap;
+    customComponents: CustomComponentContextType;
 }) {
     const value = useMemo(() => ({ ...customComponents }), [customComponents]);
